error resolved in master

+++ restservice/leavemanager/src/main/java/com/hexaware/ftp01/model/Employee.java
@@ -274,8 +274,8 @@ public class Employee {
     System.out.println("Your available leave balance was : " + empLeaveBalance);
     int empID = e.getEmpId();
     SimpleDateFormat myFormat = new SimpleDateFormat("yyyy/MM/dd");
-    int epochstartDate = myFormat.parse(date1).getTime() / 1000;
-    int epochendDate = myFormat.parse(date2).getTime() / 1000;
+    long epochstartDate = myFormat.parse(date1).getTime() / 1000;
+    long epochendDate = myFormat.parse(date2).getTime() / 1000;
     Date levAppliedOn = new Date();
     if (levNumberOfDays < 0) {
       throw new IllegalArgumentException("Enter positive value for number of days.");
